apiservice/pyproject.toml
[tool.poetry]
name = "apiservice"
version = "0.1.0"
description = ""
authors = ["Prawal Gangwar <prawal@secureailabs.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
fastapi = "^0.89.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
python-jose = "^3.3.0"
passlib = "^1.7.4"
motor = "^3.1.1"
pymongo = "^4.3.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^22.12.0"
isort = "^5.11.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
